refac: 태스크 API 및 데이터 흐름 리팩토링
