+++ src/services/taskApi.ts
@@ -1,125 +1,105 @@
-// src/services/taskApi.ts
+// src/services/taskService.ts (최종 완성본)
 
-// --- 타입 정의 ---
-type TaskStatus = 'TODO' | 'PROCESSING' | 'DONE';
-type TaskPriority = 'HIGH' | 'MEDIUM' | 'LOW';
+import axiosInstance from '@/lib/axiosInstance';
+import {
+  Task,
+  TaskListResponse,
+  TaskCreatePayload,
+  TaskUpdatePayload,
+  TaskStatus,
+} from '@/types/task';
 
-export interface TaskListResponse {
-  items: TaskResponse[];
-  summary: { todo: number; processing: number; done: number };
-  page: { number: number; size: number; totalElements: number; totalPages: number };
+/** 태스크 목록 조회를 위한 파라미터 타입 */
+export interface GetTasksParams {
+  date?: string;
+  meetingId?: number;
+  page?: number;
+  size?: number;
+  sort?: string;
+  assigneeUserId?: number;
 }
-export interface TaskResponse {
-  taskId: number;
-  teamId: number;
-  meetingId: number | null;
-  title: string;
-  description: string;
-  assigneeUserId: number | null;
-  assigneeName: string | null;
-  assigneeProfileImageUrl: string | null;
-  priorityLabel: '높음' | '보통' | '낮음';
-  statusLabel: '완료' | '진행' | '대기';
-  due: string;
-}
-export interface TaskCreateRequest {
-  title: string;
-  description?: string;
-  assigneeId: number;
-  priority?: TaskPriority;
-  status?: TaskStatus;
-  due: string;
-  meetingId?: number | null;
-}
-export type TaskUpdateRequest = Partial<Omit<TaskCreateRequest, 'assigneeId'>> & {
-  assigneeId?: number;
-};
 
-// --- API 요청 기본 설정 ---
-const API_BASE_URL = process.env.NEXT_PUBLIC_API_URL || 'https://api.dotdot.it.kr';
-const API_V1_PATH = '/api/v1';
-
-const getAuthToken = (): string | null =>
-  typeof window !== 'undefined' ? localStorage.getItem('accessToken') : null;
-const getHeaders = (): HeadersInit => {
-  const token = getAuthToken();
-  const headers: HeadersInit = { 'Content-Type': 'application/json' };
-  if (token) headers['Authorization'] = `Bearer ${token}`;
-  return headers;
-};
-
-// --- 태스크 API 요청 함수들 ---
-export const fetchTasks = async (
-  teamId: string,
-  params: {
-    startDate: string;
-    endDate?: string;
-    meetingId?: number;
-    assigneeUserId?: number;
-    page?: number;
-    size?: number;
-    sort?: string;
-  }
+/**
+ * 특정 팀의 태스크 목록을 조회합니다.
+ * @param teamId - 팀 ID
+ * @param params - 조회 조건을 담은 객체
+ * @returns 태스크 목록과 페이징 정보를 포함한 Promise
+ */
+export const getTasks = async (
+  teamId: string | number,
+  params: GetTasksParams
 ): Promise<TaskListResponse> => {
-  const queryParams = new URLSearchParams();
-  if (params.startDate) queryParams.append('startDate', params.startDate);
-  if (params.endDate) queryParams.append('endDate', params.endDate);
-  queryParams.append('page', params.page?.toString() || '0');
-  queryParams.append('size', params.size?.toString() || '100');
-  queryParams.append('sort', params.sort || 'status,asc');
-  if (params.meetingId) queryParams.append('meetingId', params.meetingId.toString());
-  if (params.assigneeUserId) queryParams.append('assigneeUserId', params.assigneeUserId.toString());
-
-  const response = await fetch(
-    `${API_BASE_URL}${API_V1_PATH}/teams/${teamId}/tasks?${queryParams.toString()}`,
-    {
-      method: 'GET',
-      headers: getHeaders(),
-    }
-  );
-  if (!response.ok) throw new Error('태스크 목록을 불러오는 데 실패했습니다.');
-  const result = await response.json();
-  return result.data;
+  try {
+    const response = await axiosInstance.get(`/api/v1/teams/${teamId}/tasks`, { params });
+    return response.data.data;
+  } catch (error) {
+    console.error(`Error fetching tasks for team ${teamId}:`, error);
+    return {
+      items: [],
+      summary: { todo: 0, processing: 0, done: 0 },
+      page: { number: 0, size: 10, totalElements: 0, totalPages: 0 },
+    };
+  }
 };
 
-export const createTask = async (teamId: string, taskData: TaskCreateRequest): Promise<number> => {
-  const response = await fetch(`${API_BASE_URL}${API_V1_PATH}/teams/${teamId}/tasks`, {
-    method: 'POST',
-    headers: getHeaders(),
-    body: JSON.stringify(taskData),
-  });
-  if (!response.ok) throw new Error('태스크 생성에 실패했습니다.');
-  const result = await response.json();
-  return result.data;
+/**
+ * 새로운 태스크를 생성합니다.
+ * @param teamId - 태스크를 생성할 팀 ID
+ * @param payload - 생성할 태스크의 데이터
+ * @returns 생성된 태스크 객체를 포함한 Promise
+ */
+export const createTask = async (
+  teamId: string | number,
+  payload: TaskCreatePayload
+): Promise<Task> => {
+  try {
+    const response = await axiosInstance.post(`/api/v1/teams/${teamId}/tasks`, payload);
+    return response.data.data;
+  } catch (error) {
+    console.error(`Error creating task for team ${teamId}:`, error);
+    throw error;
+  }
 };
 
-export const updateTask = async (
-  taskId: number,
-  taskData: TaskUpdateRequest
-): Promise<TaskResponse> => {
-  const response = await fetch(`${API_BASE_URL}${API_V1_PATH}/tasks/${taskId}`, {
-    method: 'PATCH',
-    headers: getHeaders(),
-    body: JSON.stringify(taskData),
-  });
-  if (!response.ok) throw new Error('태스크 수정에 실패했습니다.');
-  const result = await response.json();
-  return result.data;
+/**
+ * 기존 태스크의 정보를 수정합니다.
+ * @param taskId - 수정할 태스크의 ID
+ * @param payload - 수정할 태스크의 데이터
+ * @returns 수정된 태스크 객체를 포함한 Promise
+ */
+export const updateTask = async (taskId: number, payload: TaskUpdatePayload): Promise<Task> => {
+  try {
+    const response = await axiosInstance.patch(`/api/v1/tasks/${taskId}`, payload);
+    return response.data.data;
+  } catch (error) {
+    console.error(`Error updating task ${taskId}:`, error);
+    throw error;
+  }
 };
 
+/**
+ * 특정 태스크를 삭제합니다.
+ * @param taskId - 삭제할 태스크의 ID
+ */
 export const deleteTask = async (taskId: number): Promise<void> => {
-  const response = await fetch(`${API_BASE_URL}${API_V1_PATH}/tasks/${taskId}`, {
-    method: 'DELETE',
-    headers: getHeaders(),
-  });
-  if (!response.ok) throw new Error('태스크 삭제에 실패했습니다.');
+  try {
+    await axiosInstance.delete(`/api/v1/tasks/${taskId}`);
+  } catch (error) {
+    console.error(`Error deleting task ${taskId}:`, error);
+    throw error;
+  }
 };
 
-export const changeTaskStatus = async (taskId: number, status: TaskStatus): Promise<void> => {
-  const response = await fetch(`${API_BASE_URL}${API_V1_PATH}/tasks/${taskId}/status`, {
-    method: 'PATCH',
-    headers: getHeaders(),
-    body: JSON.stringify({ status }),
-  });
-  if (!response.ok) throw new Error('태스크 상태 변경에 실패했습니다.');
+/**
+ * 태스크의 상태(대기, 진행, 완료)를 변경합니다.
+ * @param taskId - 상태를 변경할 태스크의 ID
+ * @param status - 새로운 상태
+ */
+export const updateTaskStatus = async (taskId: number, status: TaskStatus): Promise<void> => {
+  try {
+    await axiosInstance.patch(`/api/v1/tasks/${taskId}/status`, { status });
+  } catch (error) {
+    console.error(`Error changing task status for ${taskId}:`, error);
+    throw error;
+  }
 };
